+++ packages/gamut-labs/src/GlobalHeader/types.tsx
@@ -5,12 +5,7 @@ import {
   AppHeaderItemWithHref,
 } from '../AppHeader/AppHeaderElements/types';
 import { AppHeaderSearch } from '../AppHeader/Search/useHeaderSearch';
-import { CrossDeviceBookmarkParts } from '../Bookmarks/types';
 import { AppHeaderNotificationSettings } from '../Notifications/types';
-
-type RenderFavorites = {
-  desktop: () => ReactNode;
-};
 
 type RenderProfile = { desktop: () => ReactNode; mobile: () => ReactNode };
 
@@ -22,13 +17,11 @@ type BaseHeader = {
   className?: string;
   hidePricing?: boolean;
   search: AppHeaderSearch;
-  crossDeviceBookmarkParts?: CrossDeviceBookmarkParts;
 };
 
 export type User = {
   avatar: string;
   displayName: string;
-  geo?: string;
   isAdmin?: boolean;
   isBusinessAdmin?: boolean;
   isBusinessSsoUser?: boolean;
@@ -45,7 +38,6 @@ export type User = {
 type LoggedInHeader = BaseHeader & {
   notifications: AppHeaderNotificationSettings;
   renderProfile?: RenderProfile;
-  renderFavorites?: RenderFavorites;
   user: User;
 };
 
@@ -73,7 +65,6 @@ export type LoadingHeader = BaseHeader & {
 
 export enum CrossDeviceItemId {
   NOTIFICATIONS = 'notifications',
-  BOOKMARKS = 'bookmarks',
   UNSET = '',
 }